pyproject.toml
[project]
name = "ecjtu-wechat-api"
version = "1.1.0"
description = "华东交通大学教务系统微信版 API"
authors = [
    { name = "mochenya" },
]
readme = "README.md"

license = "MIT"
license-files = ["LICEN[CS]E*"]

requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.100.0",
    "httpx>=0.24.0",
    "uvicorn>=0.22.0",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "twine>=6.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ecjtu_wechat_api"]

[project.urls]
Homepage = "https://github.com/mochenya/ecjtu-wechat-api"
Issues = "https://github.com/mochenya/ecjtu-wechat-api/issues"
